refactored UserInfoContainer completely
